requirements.txt
absl-py==1.4.0
altair==4.2.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==22.2.0
black==22.12.0
blinker==1.5
cachetools==5.2.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dataclasses==0.6
decorator==5.1.1
entrypoints==0.4
fire==0.5.0
flatbuffers==23.1.4
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.3
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
hydra-core==1.3.1
idna==3.4
imageio==2.24.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==6.0.0
iopath==0.1.10
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labelme2coco==0.2.1
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
mypy-extensions==0.4.3
networkx==3.0
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathspec==0.10.3
Pillow==9.3.0
pixellib==0.7.1
platformdirs==2.6.2
portalocker==2.7.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycocotools==2.0.6
pydeck==0.8.0
pydot==1.4.2
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.2.0
rsa==4.9
sahi==0.11.11
scikit-image==0.19.3
scipy==1.10.0
semver==2.13.0
shapely==2.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
streamlit==1.17.0
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminaltables==3.1.10
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.10.10
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.2
tqdm==4.64.1
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
xmljson==0.2.1
yacs==0.1.8
zipp==3.11.0
